+++ packages/tui/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [Unreleased]
+
 ## [0.56.3] - 2026-03-06
 
 ### Added