+++ packages/tui/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [0.57.0] - 2026-03-07
 
 ### Added